Use react-fontawesome instead of using scripts
This commit was also treated as a light refactor.
- Implment proper file structure
- Fix spacing
- Clean code
