+++ src/components/index.js
@@ -1,0 +1,5 @@
+import CuHackingLogo from './cuHackingLogo'
+import ParticleBackground from './particleBackground';
+import SocialLinks from './socialLinks';
+
+export {CuHackingLogo, ParticleBackground, SocialLinks};
