+++ src/components/socialLinks.module.css
@@ -1,0 +1,15 @@
+.container {
+  display: flex; 
+  flex-direction: row; 
+  justify-content: flex-start;  
+}
+
+.socialButton {
+  margin-left: 15px;
+  color: var(--textColour);
+}
+
+.socialButton:hover {
+  color: var(--primaryColour); 
+  transition: 0.2s; 
+}
