--- src/cuHackingLogo.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import cuHackingLogo from './images/logo.svg'
-
-const style = { 
-    width: "45px", 
-    height: "45px"
-}
-
-export default () => ( 
-    <img style={style} id="logo" src={cuHackingLogo} alt="Logo for cuHacking."></img>
-);
