--- src/socialLinks.js
@@ -1,27 +1,0 @@
-import React from 'react';
-
-export default () => (
-  <div>
-
-    <a class="socialwrap" target="_blank"  rel="noopener noreferrer" href="https://www.facebook.com/cuhacking/">
-      <i class="fab fa-facebook-square social"></i> 
-    </a> 
-
-    <a class="socialwrap" target="_blank"  rel="noopener noreferrer" href="https://www.linkedin.com/company/cuhacking/">  
-      <i class="fab fa-linkedin social"></i>
-    </a>
-
-    <a class="socialwrap" target="_blank"  rel="noopener noreferrer" href="https://www.instagram.com/cuHacking/">  
-      <i class="fab fa-instagram social"></i> 
-    </a> 
-
-    <a class="socialwrap" target="_blank"  rel="noopener noreferrer" href="https://twitter.com/cuhacking?lang=en">
-      <i class="fab fa-twitter social"></i>    
-    </a>        
-
-    <a class="socialwrap" target="_blank"  rel="noopener noreferrer" href="https://github.com/cuhacking">  
-      <i class="fab fa-github social"></i>
-    </a> 
-
-  </div>
-);
